BOP-66対応 金額（原価、販売価格、割引額）について、内部データの単位をパイサとし、整数値で扱うように修正（誤差の解消のため） それに伴い、表示の際に単位系を変更するように処理を修正。 また、内部DBスキーマを変更。

+++ AndroidPOS/src/com/ricoh/pos/data/OrderUpdateInfo.java
@@ -1,25 +1,25 @@
 package com.ricoh.pos.data;
 
 public class OrderUpdateInfo {
-	private double totalAmountBeforeDiscount;
-	private double discountAmount;
-	private double totalAmountAfterDiscount;
+	private long totalAmountBeforeDiscount;	// 割引前の総額。単位パイサ。
+	private long discountAmount;				// 割引額。単位パイサ。
+	private long totalAmountAfterDiscount;		// 割引後の総額。単位パイサ。
 	
-	public OrderUpdateInfo(double totalAmountBeforeDiscount, double discountAmount, double totalAmountAfterDiscount) {
+	public OrderUpdateInfo(long totalAmountBeforeDiscount, long discountAmount, long totalAmountAfterDiscount) {
 		this.totalAmountBeforeDiscount = totalAmountBeforeDiscount;
 		this.discountAmount = discountAmount;
 		this.totalAmountAfterDiscount = totalAmountAfterDiscount;
 	}
 	
-	public double getTotalAmountBeforeDiscount() {
+	public long getTotalAmountBeforeDiscount() {
 		return totalAmountBeforeDiscount;
 	}
 	
-	public double getDiscountAmount() {
+	public long getDiscountAmount() {
 		return discountAmount;
 	}
 	
-	public double getTotalAmountAfterDiscount() {
+	public long getTotalAmountAfterDiscount() {
 		return totalAmountAfterDiscount;
 	}
 }
